[Tests] Created ImageResourceTests class. Moved EXIF test there
